requirements.txt
pybit
python-dotenv
pandas
numpy
requests
loguru
PyYAML
